+++ pydsa/__init__.py
@@ -1,13 +1,13 @@
 from .bst import BSTNode
-from .selection_sort import selection_sort
-from .bubble_sort import bubble_sort
-from .insertion_sort import insertion_sort
-from .merge_sort import merge_sort
-from .quick_sort import quick_sort
-from .sleep_sort import sleep_sort
+from sorting.selection_sort import selection_sort
+from sorting.bubble_sort import bubble_sort
+from sorting.insertion_sort import insertion_sort
+from sorting.merge_sort import merge_sort
+from sorting.quick_sort import quick_sort
+from sorting.sleep_sort import sleep_sort
 from .bfs import bfs
 from .binary_search import binary_search
 from .linear_search import linear_search
 from .queue import queue
-from .counting_sort import counting_sort
+from sorting.counting_sort import counting_sort
 from .stack import Stack
