+++ pydsa/tests/test_quick_sort.py
@@ -1,4 +1,4 @@
-from pydsa.quick_sort import quick_sort
+from pydsa.sorting.quick_sort import quick_sort
 from random import randint
